Modularized gitTree and nodeModal components into gitTree folder
