+++ src/gitTree/nodeModal.html
@@ -16,7 +16,7 @@
     <p>Event: <span class="node-event"></span><p>
   </div> -->
 
-  <table id="nodeWindow">
+  <table id="nodeModal">
     <tr>
       <td>ID:</td>
       <td class="node-id"></td>
@@ -40,5 +40,5 @@
   </table>
 </body>
 
-<script type="text/javascript" src="./nodeWindow.js"></script>
+<script type="text/javascript" src="./nodeModal.js"></script>
 </html>
